refactor: changes components to utilise server-side rendering as much as possible
