+++ app/components/ShowCard.tsx
@@ -1,6 +1,5 @@
-import React, { useState } from "react";
+import React from "react";
 import { ShowStats } from "../interfaces/interfaces";
-import { updateCurrEp } from "../utils/dbFunctions";
 import AddEp from "./AddEp";
 
 const ShowCard = ({
@@ -12,8 +11,6 @@ const ShowCard = ({
   rating,
   started_watching,
 }: ShowStats) => {
-  const [epNum, setEpNum] = useState<number>(current_episode);
-
   const formattedDate = new Date(
     started_watching.seconds * 1000
   ).toDateString();
@@ -23,16 +20,11 @@ const ShowCard = ({
       <h1>{title}</h1>
       <p>{status}</p>
       <p>Rating: {rating}</p>
-      <p>
-        {epNum}/{total_episodes}
-      </p>
-      {epNum !== total_episodes ? (
-        <AddEp
-          id={id}
-          epNum={epNum}
-          setEpNum={setEpNum}
-        />
-      ) : null}
+      <AddEp
+        id={id}
+        current_episode={current_episode}
+        total_episodes={total_episodes}
+      />
       <p>{formattedDate}</p>
     </div>
   );
